+++ app/controllers/profiles_controller.rb
@@ -207,7 +207,7 @@ class ProfilesController < ApplicationController
         end
 
         allowed = [
-            :name, :audit_cookies, :audit_cookies_extensively, :audit_forms,
+            :name, :audit_cookies, :audit_nested_cookies, :audit_cookies_extensively, :audit_forms,
             :audit_headers, :audit_links, :audit_jsons, :audit_xmls,
             :audit_ui_forms,:audit_ui_inputs, :authorized_by, :scope_auto_redundant_paths,
             :audit_exclude_vector_patterns, :audit_include_vector_patterns,