+++ db/migrate/20170619155515_add_audit_nested_cookies_to_profile.rb
@@ -1,0 +1,5 @@
+class AddAuditNestedCookiesToProfile < ActiveRecord::Migration
+  def change
+    add_column :profiles, :audit_nested_cookies, :boolean
+  end
+end